backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-environ==0.11.2
sqlparse==0.4.4
pillow==10.1.0
pytils==0.4.1
djangorestframework==3.14.0
drf-extra-fields==3.7.0
